refactor(rust/sedona-raster-functions): width/height and RS_Width/RS_Height return i64 (#932)

+++ rust/sedona-raster-functions/src/rs_convexhull.rs
@@ -107,8 +107,8 @@ impl SedonaScalarKernel for RsConvexHull {
 /// of the raster in world coordinates. Due to skew/rotation in the affine
 /// transformation, each corner must be computed individually.
 fn write_convexhull_wkb(raster: &dyn RasterRef, out: &mut impl std::io::Write) -> Result<()> {
-    let width = raster.metadata().width() as i64;
-    let height = raster.metadata().height() as i64;
+    let width = raster.metadata().width();
+    let height = raster.metadata().height();
 
     // Compute the four corners in pixel coordinates:
     // Upper-left (0, 0), Upper-right (width, 0), Lower-right (width, height), Lower-left (0, height)
